requirements.txt
beautifulsoup4==4.6.0
celery==3.1.18
Django==2.0.2
djangorestframework==3.7.7
lxml==4.1.1
redis==2.10.6
requests==2.18.4
